+++ packages/ts.runtime/src/Args.ts
@@ -1,0 +1,13 @@
+export interface Args {
+  getArgs(): string[];
+  startsOnActualArguments(): boolean;
+}
+
+export const args: Args = {
+  getArgs() {
+    return process.argv;
+  },
+  startsOnActualArguments() {
+    return false;
+  },
+};