+++ packages/ts.runtime/src/Id.ts
@@ -1,0 +1,11 @@
+import { randomUUID } from "crypto";
+
+export interface Id {
+  generateId: () => string;
+}
+
+export const id: Id = {
+  generateId: function (): string {
+    return randomUUID();
+  },
+};