+++ packages/ts.runtime/src/lib/visit/visitAnyKeyword.ts
@@ -1,7 +1,7 @@
 import ts from "typescript";
 import { Visiter } from "../helpers/types";
 
-export const visitAnyKeyword: Visiter = (node, metadata) => {
+export const visitAnyKeyword: Visiter = ({ node, metadata }) => {
   return ts.factory.createObjectLiteralExpression([
     ...(metadata ?? []),
     ts.factory.createPropertyAssignment(