+++ packages/ts.runtime/src/lib/visit/visitNumberKeyword.ts
@@ -1,7 +1,7 @@
 import ts from "typescript";
 import { Visiter } from "../helpers/types";
 
-export const visitNumberKeyword: Visiter = (node, metadata) => {
+export const visitNumberKeyword: Visiter = ({node, metadata}) => {
   return ts.factory.createObjectLiteralExpression([
     ...(metadata ?? []),
     ts.factory.createPropertyAssignment(