+++ packages/ts.runtime/src/lib/visit/visitTypeParameter.ts
@@ -2,10 +2,11 @@ import ts from "typescript";
 import { Visiter } from "../helpers/types";
 import { visit } from "./visit";
 
-export const visitTypeParameter: Visiter<ts.TypeParameterDeclaration> = (
+export const visitTypeParameter: Visiter<ts.TypeParameterDeclaration> = ({
   node,
-  metadata
-) => {
+  metadata,
+  deps,
+}) => {
   return ts.factory.createObjectLiteralExpression(
     [
       ts.factory.createPropertyAssignment(
@@ -14,10 +15,12 @@ export const visitTypeParameter: Visiter<ts.TypeParameterDeclaration> = (
       ),
       ts.factory.createPropertyAssignment(
         "extends",
-        visit(
-          node.constraint ??
-            ts.factory.createKeywordTypeNode(ts.SyntaxKind.AnyKeyword)
-        ) as ts.Expression
+        visit({
+          node:
+            node.constraint ??
+            ts.factory.createKeywordTypeNode(ts.SyntaxKind.AnyKeyword),
+          deps,
+        }) as ts.Expression
       ),
     ],
     true